Refactor tournament management classes and add documentation
Improved structure and clarity of the `AddTournamentAdapter`,
`AddTournamentPresenter`, `IAddTournamentView`, and
`TournamentViewModel` classes. Added XML documentation comments
to clarify their purpose and functionality. Retained the
`IAdapter` interface to maintain the adapter pattern. Enhanced
method and property documentation for better maintainability
and readability, ensuring existing functionality is preserved.
